Fix Links.from_vector when "spiralling" in 2x2
Handle the following special case which much trouble was made to handle in other
parts of the process but not here (because I'm stupid).

When we've got a 2xN or Nx2 system (N >= 2) we can "spiral" around the Z axis to
reach places which normally wouldn't be accessible. Below shows the two examples
where we can make a "diagonal hop" in the direction not usually allowed by
"spiralling around" on the Z axis.

    (x+1, 0) <-> (x+0, 1)        (1, y+0) <-> (0, y+1)
              /                        |   |   |
        --+--/+---+--                  +---+---+
          | . |   |                    | . |   |/
        --+---+---+--                  /---+---/
          |   | . |                   /|   | . |
        --+---+/--+--                  +---+---+
              /                        |   |   |