Add namespace support to Bitfields
As proposed in #94, allows multiple fields in a BitField to share the same name
so long as they are mutually exclusive.

This commit fundamentally changes the underlying datastructure in the BitField
class from a somewhat ad-hoc dictionary-with-dependency-list form into a true
tree format. This change alone is a "good thing" (TM) and has also simplified
implementation in some places.