Fix function signatures when using contexts
Adds a new decorator which adds a rendering of a function's type to the top of a
wrapper function's docstring. This line is used by Sphinx autodoc to generate
the function signature in the documentation. This can also be used on Python 2
to work-around the fact that functools.wrapper doesn't preserve the underlying
function's prototype.