Clearer error message on SDRAM allocation failure
The error message returned when allocating SDRAM fails isn't clear that
the cause can be either that there is insufficient memory OR that the
tag assigned to the allocation is already in use.

This is now changed so that when insufficient memory is definitely the
cause (only when tag == 0) the tag is not mentioned, but that the tag is
mentioned when it may be the problem.