*EXPERIMENTAL* Proof-of-principle annealer.
Initial Commit
==============

This code is very, *very* rough and is simply a proof-of-principle
implementation of a simulated-annealing placer based roughly on the design used
in early versions of VPR.

The code is not ready for use under any circumstance and wilfully omits a huge
number of features required for real-world use.

Preliminary results
-------------------

Initial results have been highly encouraging. For a 1000 node random graph with
5-10 node fanout, HPWL-style measurements indicate that an over 80% net-length
reduction can be achieved in a minute or so. With the introduction of
windowing, this may well improve both in terms of HPWL reduction and also
runtime.

Whether this is enough (or unnecessary!) is yet to be determined by experiments
which hopefully will be coming soon to a SpiNNaker array near me...

The future
----------

Gut instinct is that for small-scale systems of a board or so, even a really
poorly optimised Python implementation would be sufficient (e.g. one as bad as
this). If a graph clustering pre-processing step could be implemented
sufficiently efficiently in Python, the same annealing system may even prove
adequate up to a modest number of 24-board frames of boards.

Moving to a C implementation could be expected to bump up the potential system
size to possibly one or two cabinets, assuming whatever clustering approach
continues to be sufficient.

Some research should be done into parallel graph clustering techniques to see
how much promise there is in this area. My hope would be "plenty" since Google
and Facepage both have large interests in this space and a similar fetish for
large distributed systems. My hope is that there is potential for a SpiNNaker
implementation of such algorithms which could then be followed by localised
(i.e. embarrassingly parallel) annealing passes.

Squashed subsequent commits add:
================================

* Add window size control as per VPR.
* Add graphviz plotting.
* Snapshot prior to "making this real"
* First nominally (rig) feature-complete SA placer.
* Remove slow min/max functions.
* Add test for netlist iteration.
* Add equality test for Machine objects.