Use core count to speed application loading
The current implementation of `load_application` queries the state of
each loaded core to ensure that loading was successful. This is good for
cases where the application map isn't _all_ the applications that will
be loaded (i.e., a count of all cores in the `wait` state is not
necessarily a count of all the cores we just tried to load), or when a
failure mode may allow the right number of cores in the wrong place.
However, in the majority of cases this is overkill because:

 1) This failure mode isn't currently possible.
 2) Most of the time `load_application` is called once to load all
    applications.

The result is that loading applications takes noticeably more time than
loading application data.

This commit provides a new parameter to `load_application`. If
`all_cores` is True then the core count is used to quickly determine
whether all applications are loaded. If False (the default) then the
current method is used.