Make encoding routes in RoutingTrees non-lossy.
Previously, the RoutingTree data structure did not explicitly encode the links
to use to get between each hop of the route. In some circumstances (e.g. NxM
with N or M <= 2), this ommision means that the structure ambiguously expresses
the route produced by the routing algorithm.

This commit (backward-incompatibly) modifies the RoutingTree data structure to
include the above informtation as well as generally make it a little cleaner.

This change should not backward-incompatibly change the rest of the
place-and-route infrastructure.