Set build timestamp time-zone to UTC.
Fixes #157.

Though I strongly suspect that the build times are actually relative to the
Europe/London timezone (i.e. what ST has on his machine), it really should be
UTC. Further, the error of a maximum of 1-ish hour is unlikely to cause
confusion since software builds are infrequent.

A very good spot by @mundya and proof that I am yet another victim of the many
[Falsehoods programmers believe about time](http://infiniteundo.com/post/25326999628/falsehoods-programmers-believe-about-time).