Remove __slots__ from user-facing modules.
The major motivation for this change is that in Python 2 you cannot pickle
objects with __slots__ in the general case. This is problematic if you wish to
dump place/route artefacts, e.g. for the purposes of plotting.

Since I don't believe we're currently memory-bound adding __slots__ probaly
isn't very useful and so remvoing this seems like a good idea.