Add work-around for BMP power-on bug
When a power-on command is sent to a BMP on a board other than board 0, the
board does not return an acknowledgement. Though this is acknowledged as a bug
in the BMP firmware by ST, there are no plans to fix it since in practical
systems there are no scenarios where board 0 does not exist.

The fix for this is always sending power-on/off commands to board 0, whether
board 0 is in the mask or not.