Make full-frame power-on the default for rig-power
This commit makes powering on all boards in a frame the default behaviour.

Note that this works since if the power-on/off mask contains at least one valid
board, the command does not fail, even if the mask refers to a board which is
not present. As a result, this command still works for single-board setups.