Add Machine 'issubset' method
This test allows you to verify that one machine offers a (non-strict) subset of
the resources of another. This is useful, for example, if you have a stored
placement solution which you wish to use on a larger machine (or one a machine
whose dead chips have been subsequently replaced).