Add SameChipConstraint place-and-route constraint.
This constraint allows applications to specify that a given collection of
vertices should always be placed on the same chip. This functionality is being
added to address an iminent need for this functionality in nengo_spinnaker and
a perhaps less iminent need in pinus_rigida.

The functionality has been retrofitted to the existing suite of placement
algorithms by defining a pair of utility functions which temporarily squash
same-chip-constrained vertices together into a single vertex. Placement may now
proceed unchanged. Once placed, the squashed vertices are re-expanded into its
constituent vertices and the resulting placement can be returned.

The constraint has no effect on the allocation and routing stages of the
toolchain and so these algorithms remain unchanged.