Fix coverage of docstrings utility module.
The enum34 library sets the docstring of enums to None by default while native
Python>=3.4 enums have a default docstring. As such some of the code to handle
this situation becomes dead in Python 3.4 onwards when the enum34 library is
not used. This commit thus disables coverage checking for the code since it
will not get meaningfully hit in Python 3.4 onwards.