Refactor hilbert placer.
This change extracts the kernel of the sequential hilbert placement algorithm
and makes it available as a standalone 'sequential' placer which can be
supplied with a user-defined vertex and chip use ordering. The breadth-first
vertex ordering heuristic is also broken out (and for completeness/example
usage a breadth-first sequential placement algorithm is provided).

In addition to being a generally cleaner design this will specifically make it
much easier to produce simple application-specific sequential placement
algorithms. For example when mapping a 2D grid-style application a raster-scan
sequential placement is often pretty-much what you'd want and this will make
such an algorithm very simple to implement (coming soon!).