Make MachineController.boot smarter.
This commit makes two changes to the MachineController.boot() method:

* It (optionally) checks to see if the machine is already booted and does
  nothing if it is. This makes calling the method as a routine part of an
  application will no longer incur a significant runtime penalty when the
  machine is already booted. When this check is enabled it will also cause an
  error if you attempt to boot a BMP.
* It (optionally) checks that the machine responds to sver after it was booted.

Both optional checks are enabled by default. Though this is technically
breaking backward compatibility for anyone relying on the boot method not
sending any SCP commands, this behaviour should be sufficiently niche that this
can be considered a non-breaking change.

This commit also simplifies the rig-boot command's implementation and includes
updated documentation.

Incorporates a number of contributions and clarifications by @mundya.