Add (preliminary) get_idle_cores and get_resource_constraints to MachineController.
This commit adds two new methods to MachineController motivated in issue #202
(Use only unused cores).

In brief, the get_resource_constraints method returns a set of place-and-route
constraints which can be supplied to Rig's place-and-route library functions
which prevents non-idle cores being used. Due to a lack of support from the
current version of SC&MP, used SDRAM is not constrained but the API is ready
for this to be added in the near future...

To facilitate implementation of get_resource_constraints, the new
get_idle_cores method provides a list of cores in the Idle state on a given
chip. This preliminary implementation simply reads each core's VCPU struct.
Though this works, it is obnoxiously inefficient and very slow on high-latency
connections. The intent is that a future version of SC&MP will provide this
information as a bitmap or other quickly read value, making this funciton much
faster.