Improve documentation for sdram_alloc[_as_filelike].
This commit adds documentation which explains how the `tag` feature is used,
including noting the SARK function to use to retrieve tagged SDRAM addresses.

In addition the doctest example in the MemoryIO docstring was corrected to
include seeking and include the return values from write() calls. A brief
demonstration of MemoryIO was also added to the sdram_alloc_as_filelike
documentation to entice new users.