Add `place_and_route_wrapper`, deprecating `wrapper`.
This commit deprecates `rig.place_and_route.wrapper` in favour of
`rig.place_and_route.place_and_route_wrapper`. This is done for two reasons:

1. The old name was ambiguous when imported directly as in many applications.

2. Manually adding ReserveResourceConstraints etc. is now nolonger recommended
   practice following the availability of
   `MachineController.get_resource_constraints`. The new function no longer
   includes automatic insertion of monitor-reserving constraints as a result.

The old `wrapper` function remains available for backward compatibility with
older applications but should not be used in new code. Users of the old
function are recommended to migrate to the new place_and_route_wrapper function
to gain the benefits of automatically avoiding any non-idle core, not just
monitors.