Test that the stop signal now works.
This test actually found a bug in the "fixed" SC&MP implementation whereby the
first allocation on a chip with no applications would be freed correctly, but
not the second.

This was caused by a "clean" flag being used to skip the "free" process when it
wasn't needed. This clean flag was cleared on startup and on app load but not
upon tag/SDRAM allocation. This bug has now been repaired.

The new SC&MP included in this release is compiled from commit 1392251 in:

https://bitbucket.org/mossblaser/spinnaker_tools_134/commits/1392251effded9541978aa8f75c1a5475b06af88

This version of SC&MP also includes patches for app loading and getting chip
info from 6d64c80 and 8d638a0 in rig respectively along with the
partially-working app-stop fix from a60361e.