Add get_system_info and SystemInfo.
The new MachineController.get_system_info() and the SystemInfo object returned
are intended to replace get_machine() and the Machine object for the purposes
of establishing the status, connectivity and general health of a SpiNNaker
machine. The SystemInfo object features many of the same tests and iterators as
Machine objects making it a drop-in replacement for many applications. Unlike
the Machine object, the SystemInfo object is explicitly aware of cores (and
their app-states) and correspondingly offers flexible support for iterating
over and querying them.

As a result of this change, the `get_machine` is now deprecated and
the preferred way of getting a `Machine` for P&R purposes is via a new utility
function which converts a SystemInfo into a machine.

The experimental get_resource_constraints method has also been removed in
favour of a new utility function which converts SystemInfo objects into lists
of constraints.