Use get_system_info for app_stop work-around.
Previously used get_machine which meant that get_machine may sometimes get
called in this workaround before the test for the get_machine deprecation
warning. Since the warning is only emitted once this makes the test that
ensures the warning is produced fail non-deterministically depending on the
ordering of the tests.