Insert entries above blocks of similar entries
Insert entries above blocks of entries of equivalent generality. This
makes that the Python implementation of OC matches (a) the C version for
SpiNNaker (b) that described in the HPSR submission.

For table ordering, see
https://github.com/project-rig/rig_routing_tables/blob/master/include/ordered_covering.h#L16-L57

For downcheck preferring to modify more significant bits in merges see
https://github.com/project-rig/rig_routing_tables/blob/master/include/ordered_covering.h#L285-L292

For merge application, see
https://github.com/project-rig/rig_routing_tables/blob/master/include/ordered_covering.h#L409-L485