Improve NER neighbourhood search performance.
This commit reduces the router runtime to about 3/5 of its old runtime in
situations where long routes must be generated.

When generating routing trees, the NER algorithm searches for nearby vertices
(or pieces of route) to route to rather than always routing back to the source
of the net. Originally this was achieved by searching in a spiral pattern
outward from the destination vertex looking for pieces of route to attach to.
With the default search radius of 20 this requires 1261 checks, all of which
are carried out when the destination is more than 20 hops from another part of
the route.

A new, alternative implementation instead iterates over every known route
segment to find out if any are within the specified search radius. In practice
this often results in substantially fewer checks being required. To reap the
benefits of both approaches, a simple heuristic automatically selects the
approach to use on a route-by-route basis.

To support these changes, the mesh and torus distance calculations have been
uglified/optimised (for a measurable but small performance improvement) since
they now form part of the hot code path in NER.