Fix bug in machine boot kwargs
Some of the keyword arguments used to boot SpiNNaker machines were
overloaded. In particular, `boot_delay` is present both in the `sv`
struct and in the list of arguments expected by `boot`. This commit
fixes this issue by making it possible to set the SV overrides
explicitly.

Note that the API changes are backwards compatible, this only adds to
the commands accepted by `boot`.