Determine the cause of a failed alloc
Allocating SDRAM can fail either because there is insufficient memory
available or because the requested tag is already in use. This commit
modifies the `sdram_alloc` method so that it reads the allocation table
to determine the cause of the failure and reports this in the error
that is raised.

This should make diagnosing memory allocation failures easier.