Remove tests for signal IntEnum types
Testing of one IntEnum type value is 'in' another IntEnum type now returns true
if the matching integer appears in both. In the past this wasn't the case.
Manually making the implementation do the equivalent type checking seems
unnecessary and would be better served by a future move to a typed API.

To be clear, this change makes the test less strict by no longer checking to
see that other IntEnum types are rejected -- since they're not!